fix: removed the context variable for contract in favour of the event address (untested)
