+++ packages/subgraph/src/packages/token/governance-erc20.ts
@@ -5,61 +5,14 @@ import {
   DelegateVotesChanged,
 } from '../../../generated/templates/GovernanceERC20/GovernanceERC20';
 import {Transfer} from '../../../generated/templates/TokenVoting/ERC20';
-import {ADDRESS_ZERO} from '../../utils/constants';
 import {generateMemberEntityId} from '../../utils/ids';
+import {
+  TokenVotingMemberResult,
+  getDelegateeId,
+  getERC20Balance,
+  getVotingPower,
+} from './utils';
 import {Address, BigInt, dataSource, store} from '@graphprotocol/graph-ts';
-
-function getERC20Balance(user: Address, tokenAddress: Address): BigInt {
-  let contract = GovernanceERC20Contract.bind(tokenAddress);
-  let balance = contract.balanceOf(user);
-  return balance;
-}
-
-function getDelegation(user: Address, tokenAddress: Address): string | null {
-  let contract = GovernanceERC20Contract.bind(tokenAddress);
-  let delegate = contract.delegates(user);
-
-  return delegate === Address.fromString(ADDRESS_ZERO)
-    ? null
-    : delegate.toHexString();
-}
-
-function getDelegateeId(
-  user: Address,
-  tokenAddress: Address,
-  pluginId: string
-): string | null {
-  let delegatee = getDelegation(user, tokenAddress);
-  return delegatee
-    ? generateMemberEntityId(
-        Address.fromString(pluginId),
-        Address.fromString(user.toHexString())
-      )
-    : null;
-}
-
-function getVotingPower(user: Address, tokenAddress: Address): BigInt {
-  let contract = GovernanceERC20Contract.bind(tokenAddress);
-  let votingPower = contract.getVotes(user);
-  return votingPower;
-}
-
-/**
- * A container for the result of the `getOrCreateMember` function.
- * @param entity - The `TokenVotingMember` entity.
- * @param createdNew - A boolean indicating whether the entity was created new
- * or if false it was previously created. If the entity was created new, it already
- * has the latest balance of the user, so no need to then update the balance.
- */
-class TokenVotingMemberResult {
-  entity: TokenVotingMember;
-  createdNew: boolean;
-
-  constructor(entity: TokenVotingMember, createNew: boolean) {
-    this.entity = entity;
-    this.createdNew = createNew;
-  }
-}
 
 function getOrCreateMember(
   user: Address,
@@ -90,7 +43,7 @@ function getOrCreateMember(
 export function handleTransfer(event: Transfer): void {
   let context = dataSource.context();
   let pluginId = context.getString('pluginId');
-  let tokenAddress = Address.fromString(context.getString('tokenAddress'));
+  let tokenAddress = event.address;
 
   if (event.params.from != Address.zero()) {
     let result = getOrCreateMember(event.params.from, pluginId, tokenAddress);
@@ -118,7 +71,7 @@ export function handleTransfer(event: Transfer): void {
 export function handleDelegateChanged(event: DelegateChanged): void {
   let context = dataSource.context();
   let pluginId = context.getString('pluginId');
-  let tokenAddress = Address.fromString(context.getString('tokenAddress'));
+  let tokenAddress = event.address;
 
   const toDelegate = event.params.toDelegate;
 
@@ -167,7 +120,7 @@ export function handleDelegateVotesChanged(event: DelegateVotesChanged): void {
 
   const context = dataSource.context();
   const pluginId = context.getString('pluginId');
-  const tokenAddress = Address.fromString(context.getString('tokenAddress'));
+  const tokenAddress = event.address;
 
   let result = getOrCreateMember(delegate, pluginId, tokenAddress);
   let member = result.entity;
